+++ gearoenix/opengl/gx-gl-light.hpp
@@ -63,7 +63,7 @@ private:
         core::job::EndCaller<>&& end_callback) override;
     void build_shadow_caster_directional(
         const std::string& name,
-        std::size_t shadow_map_resolution,
+        std::uint32_t shadow_map_resolution,
         float camera_far,
         float camera_near,
         float camera_aspect,
