+++ gearoenix/render/texture/gx-rnd-txt-image.hpp
@@ -20,36 +20,36 @@ public:
     static void decode(
         platform::stream::Stream* file,
         std::vector<unsigned char>& encoded_data,
-        std::size_t& img_width,
-        std::size_t& img_height);
+        std::uint32_t& img_width,
+        std::uint32_t& img_height);
     static void decode(
         const unsigned char* formatted_data,
         std::size_t formatted_size,
         std::optional<std::size_t> requested_channels,
         std::vector<unsigned char>& decoded_data,
-        std::size_t& img_width,
-        std::size_t& img_height,
-        std::size_t& img_channels);
+        std::uint32_t& img_width,
+        std::uint32_t& img_height,
+        std::uint32_t& img_channels);
     static void decode(
         const unsigned char* formatted_data,
         std::size_t formatted_size,
         std::optional<std::size_t> requested_channels,
         std::vector<float>& decoded_data,
-        std::size_t& img_width,
-        std::size_t& img_height,
-        std::size_t& img_channels);
+        std::uint32_t& img_width,
+        std::uint32_t& img_height,
+        std::uint32_t& img_channels);
     static void encode_png(
         platform::stream::Stream& file,
         const std::uint8_t* data,
-        std::size_t img_width,
-        std::size_t img_height,
-        std::size_t components_count);
+        std::uint32_t img_width,
+        std::uint32_t img_height,
+        std::uint32_t components_count);
     static void encode_hdr(
         platform::stream::Stream& file,
         const void* data,
-        std::size_t img_width,
-        std::size_t img_height,
-        std::size_t components_count);
+        std::uint32_t img_width,
+        std::uint32_t img_height,
+        std::uint32_t components_count);
 };
 }
 #endif
